refactor: Rules Folder im Test Package erstellt, und Tests ausgelagert
